AITA-39 Restructured and remove unnecessary components
