--- src/views/components/SubmitButton/index.tsx
@@ -1,7 +1,0 @@
-import { SubmitButtonStyled } from './styled';
-
-export const SubmitButton: React.FC<{ title?: string; onClick: () => void }> = ({
-  title = 'Submit',
-}) => {
-  return <SubmitButtonStyled>{title}</SubmitButtonStyled>;
-};
